btf: add CO-RE field relocations
Field relocations record the path through a nested structure. They
are generated by the following macros:

    bpf_core_field_exists(foo->bar)
    bpf_core_field_size(foo->bar)

BPF_FIELD_BYTE_OFFSET is currently not generated directly. To resolve
such relocations we follow the accessor through the (possibly) nested
array, struct or union.

For arrays we use the accessor as the index into the array, after
checking that the index is within bounds. The bounds check is not
performed for  flexible arrays at the end of a structure or union:

    struct flex {
	int bar;
        char flex[];
    };

For structs and unions we use the accessor to look up the field in
the local type and then try to find a field with the same name in
the target type. Anonymous unions and structs are ignored for the
purpose of relocation:

    struct bar {
        int bar;
        union {
            int foo;
        };
    };

    // equivalent to

    struct bar {
        int bar;
        int foo;
    };